Fix execution speed regression in cleanup
Replace multiple conversion from PathBuf to PathAbs in
`sort_unstable_by()` and `dedup_by()` with just one BTreeMap that
relates both entries. When introduced in this collection, the data is
ordered and deduplicated. This way the program execution speed is x4
faster, almost the same that RnR had before introducing deduplication in
cleanup function.